+++ src/main/java/com/br/financeiro/service/UsuarioService.java
@@ -1,5 +1,6 @@
 package com.br.financeiro.service;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.BeanUtils;
@@ -9,6 +10,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.br.financeiro.exceptionhandler.CustomRuntimeException;
+import com.br.financeiro.model.Pessoa;
 import com.br.financeiro.model.Usuario;
 import com.br.financeiro.repository.UsuarioRepository;
 
@@ -17,9 +19,15 @@ public class UsuarioService {
 	
 	@Autowired
 	private UsuarioRepository usuarioRepository;
-
+	
+	@Autowired
+	private PessoaService pessoaService;
+	
 	public Usuario salvar(Usuario entidade) {
 		Optional<Usuario> pesquisarPorEmail = this.usuarioRepository.findByEmail(entidade.getEmail());
+		if(entidade.getPermissoes().isEmpty()) {
+			throw new CustomRuntimeException("Selecione pelos menos 1 permissão!");
+		}
 		if(pesquisarPorEmail.isPresent()) {
 			throw new CustomRuntimeException("e-mail já cadastrado!");
 		}
@@ -35,6 +43,9 @@ public class UsuarioService {
 		}
 		if(pesquisarPorEmail.isPresent() && entidadeSalva.get().getId() != entidade.getId()) {
 			throw new CustomRuntimeException("E-mail já existe!");	
+		}
+		if(entidade.getPermissoes().isEmpty()) {
+			throw new CustomRuntimeException("Selecione pelos menos 1 permissão!");
 		}
 		if(entidade.getSenha() == null) {
 			entidade.setSenha(entidadeSalva.get().getSenha());
@@ -52,12 +63,24 @@ public class UsuarioService {
 	}
 	
 	public void excluir(Long id) {
+		Optional<Usuario> entidadeSalva = this.usuarioRepository.findById(id);
+		Optional<Pessoa> entidadeAssociada = this.pessoaService.buscarUsuarioById(id);
+		if(!entidadeSalva.isPresent()) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		if(entidadeAssociada.isPresent()){
+			throw new CustomRuntimeException("Erro ao excluir pessoa, existe usuário vinculado!");	
+		}
 		this.usuarioRepository.deleteById(id);
 	}
 	
 	public Iterable<Usuario> listar() {
 		return usuarioRepository.findAll();
 	}	
+	
+	public List<Usuario> listaUsuariosDisponiveis() {
+		return this.usuarioRepository.listaUsuariosDisponiveis();
+	}
 	
 	public String gerarHash(String password) {
 		BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
